Fixed issue with Log file location
The log file name was being added to the end of the executing program
that used this class. Fixed the issue by using path to strip the
filename and extension then added in a default log file name. Also
updated the archiving feature to correctly name the file and put on a
txt extension.