Corrected Comment style, changed method name added Append switch
Changed the comments to correct format for Visual Studio methods.
Changed the name of the AddEvent method to AddLog to make it clear in
case using the EventManager at the same time as the LogManager.
Added ability to force append, although it still respects large files
and handles in the same method